week4/taxi_rides_ny/pyproject.toml
[project]
name = "taxi-rides-ny"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-duckdb>=1.10.1",
    "dbt-bigquery>=1.9.0",
    "duckdb>=1.4.4",
    "google-cloud-bigquery>=3.38.0",
]
